requirements.txt
fastapi
uvicorn[standard]

PyPDF2

faiss-cpu
python-docx

openai
httpx
numpy
python-multipart
